+++ jina/serve/runtimes/gateway/http/app.py
@@ -13,24 +13,19 @@ from jina.logging.logger import JinaLogger
 if TYPE_CHECKING:
     from prometheus_client import CollectorRegistry
 
-    from jina.serve.networking import GrpcConnectionPool
-    from jina.serve.runtimes.gateway.graph.topology_graph import TopologyGraph
-
 
 def get_fastapi_app(
-    args: 'argparse.Namespace',
-    topology_graph: 'TopologyGraph',
-    connection_pool: 'GrpcConnectionPool',
-    logger: 'JinaLogger',
-    metrics_registry: Optional['CollectorRegistry'] = None,
+        args: 'argparse.Namespace',
+        logger: 'JinaLogger',
+        timeout_send: Optional[float] = None,
+        metrics_registry: Optional['CollectorRegistry'] = None,
 ):
     """
     Get the app from FastAPI as the REST interface.
 
     :param args: passed arguments.
-    :param topology_graph: topology graph that manages the logic of sending to the proper executors.
-    :param connection_pool: Connection Pool to handle multiple replicas and sending to different of them
     :param logger: Jina logger.
+    :param timeout_send: Timeout to be used when sending to Executors
     :param metrics_registry: optional metrics registry for prometheus used if we need to expose metrics from the executor or from the data request handler
     :return: fastapi app
     """
@@ -47,8 +42,8 @@ def get_fastapi_app(
     app = FastAPI(
         title=args.title or 'My Jina Service',
         description=args.description
-        or 'This is my awesome service. You can set `title` and `description` in your `Flow` or `Gateway` '
-        'to customize the title and description.',
+                    or 'This is my awesome service. You can set `title` and `description` in your `Flow` or `Gateway` '
+                       'to customize the title and description.',
         version=__version__,
     )
 
@@ -62,24 +57,30 @@ def get_fastapi_app(
         )
         logger.warning('CORS is enabled. This service is accessible from any website!')
 
-    from jina.serve.runtimes.gateway.request_handling import RequestHandler
-    from jina.serve.stream import RequestStreamer
+    from jina.serve.bff import GatewayBFF
 
-    request_handler = RequestHandler(metrics_registry, args.name)
+    import json
 
-    streamer = RequestStreamer(
-        args=args,
-        request_handler=request_handler.handle_request(
-            graph=topology_graph, connection_pool=connection_pool
-        ),
-        result_handler=request_handler.handle_result(),
-    )
-    streamer.Call = streamer.stream
+    graph_description = json.loads(args.graph_description)
+    graph_conditions = json.loads(args.graph_conditions)
+    deployments_addresses = json.loads(args.deployments_addresses)
+    deployments_disable_reduce = json.loads(args.deployments_disable_reduce)
+
+    gateway_bff = GatewayBFF(graph_representation=graph_description,
+                             executor_addresses=deployments_addresses,
+                             graph_conditions=graph_conditions,
+                             deployments_disable_reduce=deployments_disable_reduce,
+                             timeout_send=timeout_send,
+                             retries=args.retries,
+                             compression=args.compression,
+                             runtime_name=args.name,
+                             prefetch=args.prefetch,
+                             logger=logger,
+                             metrics_registry=metrics_registry)
 
     @app.on_event('shutdown')
     async def _shutdown():
-        await streamer.wait_floating_requests_end()
-        await connection_pool.close()
+        await gateway_bff.close()
 
     openapi_tags = []
     if not args.no_debug_endpoints:
@@ -87,7 +88,7 @@ def get_fastapi_app(
             {
                 'name': 'Debug',
                 'description': 'Debugging interface. In production, you should hide them by setting '
-                '`--no-debug-endpoints` in `Flow`/`Gateway`.',
+                               '`--no-debug-endpoints` in `Flow`/`Gateway`.',
             }
         )
 
@@ -118,7 +119,7 @@ def get_fastapi_app(
         @app.get(
             path='/dry_run',
             summary='Get the readiness of Jina Flow service, sends an empty DocumentArray to the complete Flow to '
-            'validate connectivity',
+                    'validate connectivity',
             response_model=PROTO_TO_PYDANTIC_MODELS.StatusProto,
         )
         async def _flow_health():
@@ -175,7 +176,7 @@ def get_fastapi_app(
             # do not add response_model here, this debug endpoint should not restricts the response model
         )
         async def post(
-            body: JinaEndpointRequestModel, response: Response
+                body: JinaEndpointRequestModel, response: Response
         ):  # 'response' is a FastAPI response, not a Jina response
             """
             Post a data request to some endpoint.
@@ -282,7 +283,7 @@ def get_fastapi_app(
             {
                 'name': 'CRUD',
                 'description': 'CRUD interface. If your service does not implement those interfaces, you can should '
-                'hide them by setting `--no-crud-endpoints` in `Flow`/`Gateway`.',
+                               'hide them by setting `--no-crud-endpoints` in `Flow`/`Gateway`.',
             }
         )
         crud = {
@@ -321,7 +322,7 @@ def get_fastapi_app(
             from docarray import DocumentArray
 
             async def get_docs_from_endpoint(
-                data, target_executor, parameters, exec_endpoint
+                    data, target_executor, parameters, exec_endpoint
             ):
                 req_generator_input = {
                     'data': [asdict(d) for d in data],
@@ -332,8 +333,8 @@ def get_fastapi_app(
                 }
 
                 if (
-                    req_generator_input['data'] is not None
-                    and 'docs' in req_generator_input['data']
+                        req_generator_input['data'] is not None
+                        and 'docs' in req_generator_input['data']
                 ):
                     req_generator_input['data'] = req_generator_input['data']['docs']
                 try:
@@ -351,11 +352,11 @@ def get_fastapi_app(
             class Mutation:
                 @strawberry.mutation
                 async def docs(
-                    self,
-                    data: Optional[List[StrawberryDocumentInput]] = None,
-                    target_executor: Optional[str] = None,
-                    parameters: Optional[JSONScalar] = None,
-                    exec_endpoint: str = '/search',
+                        self,
+                        data: Optional[List[StrawberryDocumentInput]] = None,
+                        target_executor: Optional[str] = None,
+                        parameters: Optional[JSONScalar] = None,
+                        exec_endpoint: str = '/search',
                 ) -> List[StrawberryDocument]:
                     return await get_docs_from_endpoint(
                         data, target_executor, parameters, exec_endpoint
@@ -365,11 +366,11 @@ def get_fastapi_app(
             class Query:
                 @strawberry.field
                 async def docs(
-                    self,
-                    data: Optional[List[StrawberryDocumentInput]] = None,
-                    target_executor: Optional[str] = None,
-                    parameters: Optional[JSONScalar] = None,
-                    exec_endpoint: str = '/search',
+                        self,
+                        data: Optional[List[StrawberryDocumentInput]] = None,
+                        target_executor: Optional[str] = None,
+                        parameters: Optional[JSONScalar] = None,
+                        exec_endpoint: str = '/search',
                 ) -> List[StrawberryDocument]:
                     return await get_docs_from_endpoint(
                         data, target_executor, parameters, exec_endpoint
@@ -385,7 +386,7 @@ def get_fastapi_app(
         :param request_iterator: request iterator, with length of 1
         :return: the first result from the request iterator
         """
-        async for k in streamer.stream(request_iterator=request_iterator):
+        async for k in gateway_bff.stream(request_iterator=request_iterator):
             request_dict = k.to_dict()
             return request_dict