+++ jina/serve/runtimes/gateway/websocket/__init__.py
@@ -66,16 +66,13 @@ class WebSocketGatewayRuntime(GatewayRuntime):
                 if ssl_file not in uvicorn_kwargs.keys():
                     uvicorn_kwargs[ssl_file] = getattr(self.args, ssl_file)
 
-        self._set_topology_graph()
-        self._set_connection_pool()
         self._server = UviServer(
             config=Config(
                 app=extend_rest_interface(
                     get_fastapi_app(
-                        self.args,
-                        topology_graph=self._topology_graph,
-                        connection_pool=self._connection_pool,
+                        args=self.args,
                         logger=self.logger,
+                        timeout_send=self.timeout_send,
                         metrics_registry=self.metrics_registry,
                     )
                 ),
@@ -90,7 +87,6 @@ class WebSocketGatewayRuntime(GatewayRuntime):
 
     async def async_run_forever(self):
         """Running method of ther server."""
-        self._connection_pool.start()
         await self._server.serve()
 
     async def _wait_for_cancel(self):
@@ -104,7 +100,6 @@ class WebSocketGatewayRuntime(GatewayRuntime):
     async def async_teardown(self):
         """Shutdown the server."""
         await self._server.shutdown()
-        await self._connection_pool.close()
 
     async def async_cancel(self):
         """Stop the server."""