+++ scripts/get-openapi-schemas.py
@@ -2,8 +2,6 @@ import json
 
 from jina.logging.logger import JinaLogger
 from jina.parsers import set_gateway_parser
-from jina.serve.networking import GrpcConnectionPool
-from jina.serve.runtimes.gateway.graph.topology_graph import TopologyGraph
 from jina.serve.runtimes.gateway.http.app import get_fastapi_app
 
 JINA_LOGO_URL = 'https://api.jina.ai/logo/logo-product/jina-core/horizontal-layout/colored/Product%20logo_Core_vertical_colorful%402x-margin.png'
@@ -14,8 +12,6 @@ args = set_gateway_parser().parse_args([])
 logger = JinaLogger('')
 gateway_app = get_fastapi_app(
     args,
-    topology_graph=TopologyGraph({}),
-    connection_pool=GrpcConnectionPool(logger=logger),
     logger=logger,
 )
 gateway_schema = gateway_app.openapi()